Update and rename CoursColor.tsx to CoursColor.js
